finish header and search box
