+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/index.js":"1","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/reportWebVitals.js":"2","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/App.js":"3","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/Header/Header.js":"4"},{"size":737,"mtime":1611844925363,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":1611837481848,"results":"7","hashOfConfig":"6"},{"size":673,"mtime":1611843471754,"results":"8","hashOfConfig":"6"},{"size":385,"mtime":1611844189766,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"6am71u",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/index.js",["19"],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/reportWebVitals.js",[],["20","21"],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/App.js",["22"],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/Header/Header.js",[],{"ruleId":"23","severity":1,"message":"24","line":3,"column":8,"nodeType":"25","messageId":"26","endLine":3,"endColumn":11},{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","replacedBy":"30"},{"ruleId":"23","severity":1,"message":"31","line":1,"column":17,"nodeType":"25","messageId":"26","endLine":1,"endColumn":26},"no-unused-vars","'App' is defined but never used.","Identifier","unusedVar","no-native-reassign",["32"],"no-negated-in-lhs",["33"],"'Component' is defined but never used.","no-global-assign","no-unsafe-negation"]
+[{"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/index.js":"1","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/reportWebVitals.js":"2","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/App.js":"3","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/Header/Header.js":"4","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Pages/Index.js":"5","/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/SearchBox/SearchBox.js":"6"},{"size":539,"mtime":1611885770064,"results":"7","hashOfConfig":"8"},{"size":362,"mtime":1611837481848,"results":"9","hashOfConfig":"8"},{"size":250,"mtime":1611884445915,"results":"10","hashOfConfig":"8"},{"size":385,"mtime":1611844189766,"results":"11","hashOfConfig":"8"},{"size":183,"mtime":1611888192741,"results":"12","hashOfConfig":"8"},{"size":756,"mtime":1611888214628,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"6am71u",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/index.js",[],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/reportWebVitals.js",[],["27","28"],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/App.js",[],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/Header/Header.js",[],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Pages/Index.js",["29"],"/Users/igormagro/OneDrive/Development/challenges/join-com/coding-challenge-frontend-react/src/Components/SearchBox/SearchBox.js",[],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":1,"column":17,"nodeType":"36","messageId":"37","endLine":1,"endColumn":25},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"no-unused-vars","'Fragment' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
