+++ src/Components/Header/Styles.css
@@ -1,7 +1,6 @@
 .root {
-  position: fixed;
   height: 100px;
-  width: 100%;
+
   display: flex;
   flex-direction: row;
   justify-content: left;
@@ -14,19 +13,20 @@
   margin-right: 32px;
 }
 
-.title_box{
-  display: flex;
+.title_box {
+  display: inline-flex;
   flex-direction: column;
   height: 100px;
   justify-content: center;
 }
 
-.title_box h1,p{
+.title_box h1,
+p {
   margin: 0;
 }
-.title_box h1{
-  font-size: 4rem;
+.title_box h1 {
+  font-size: 3rem;
 }
-.title_box p{
+.title_box p {
   font-size: 2rem;
 }
