+++ src/Components/SearchBox/Styles.css
@@ -1,0 +1,65 @@
+.box {
+  display: inline-flex;
+  width: -webkit-fill-available;
+  height: 100px;
+  justify-content: left;
+  align-items: center;
+  padding: 20px;
+  position: relative;
+}
+
+form {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-end;
+}
+
+input {
+  outline: none;
+  height: 30px;
+  border-radius: 4px;
+  border: none;
+  border: 0.5px solid rgba(0, 0, 0, 0.233);
+  /* box-shadow: 1px 2px 2px rgba(0, 0, 0, 0.233); */
+  width: 100%;
+}
+
+.btn {
+  height: 30px;
+  border-radius: 4px;
+  width: 150px;
+  padding: 1px;
+  cursor: pointer;
+  border: none;
+  box-shadow: 1px 2px 2px rgba(0, 0, 0, 0.233);
+  outline: none;
+}
+.btn:hover {
+  background-color: rgb(161, 160, 160);
+  transition: all 0.8s;
+  font-weight: bold;
+}
+
+.ipt_description {
+  width: 40%;
+  margin-right: 16px;
+  padding-left: 4px;
+}
+
+.box__date {
+  flex-direction: column;
+  width: 150px;
+  margin: 0 16px 0;
+}
+
+.box__date p {
+  padding-bottom: 4px;
+  font-weight: 700;
+}
+.box__p {
+  position: absolute;
+  right: 20px;
+  bottom: 0;
+  font-weight: 600;
+}
