+++ src/Pages/Index.js
@@ -1,0 +1,8 @@
+import React, { Fragment } from "react";
+import SearchBox from "../Components/SearchBox/SearchBox";
+
+const Index = () => {
+  return <SearchBox total={20}/>;
+};
+
+export default Index;
