+++ src/index.css
@@ -1,0 +1,5 @@
+.body {
+  width: 70%;
+  min-height: 100%;
+  font-family: "Roboto Slab", serif;
+}
